Change wallet account balance to BigDecimal

+++ src/main/java/pl/sdag2/entity/Wallet.java
@@ -7,6 +7,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 @Getter
@@ -16,7 +17,7 @@ public class Wallet {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String user;
-    private int accountBalance;
+    private BigDecimal accountBalance;
     private PayMethodEnum payMethod;
 
 }
